weather_project/requirements.txt
django 
djangorestframework
opencage
drf-yasg
psycopg2
